requirements.txt
numpy
scikit-learn
torch
torchvision
pillow
matplotlib
kagglehub